address review comments 2
